// src/app/interfaces/PagamentoAtualizar.ts
export interface PagamentoAtualizar {
    dinheiro: number;
    debito: number;
    credito: number;
    creditoParcelado: number;
    parcelas: number;
    pix: number;
    desistente: number;
    dataRegistro: Date;
    receber: number;
    descontar: number;
    observacao: string;
    fichaConsumidor: number;
    voluntario: number;
    siao: number;
    idPagamento: number;
}